修改分类icon空指针bug

+++ game-admin/game-admin-controller/src/main/java/com/fulu/game/admin/controller/CategoryController.java
@@ -35,6 +35,7 @@ public class CategoryController extends BaseController {
     private SalesModeService salesModeService;
     @Autowired
     private OssUtil ossUtil;
+
     /**
      * 内容列表
      *
@@ -123,7 +124,7 @@ public class CategoryController extends BaseController {
                     tagService.update(tag);
                 }
             }
-            if(!origCategory.getIcon().equals(category.getIcon())){
+            if (origCategory.getIcon() != null && !origCategory.getIcon().equals(category.getIcon())) {
                 ossUtil.deleteFile(origCategory.getIcon());
             }
             category.setUpdateTime(new Date());
@@ -159,7 +160,7 @@ public class CategoryController extends BaseController {
                                   @RequestParam(required = true) String name,
                                   @RequestParam(required = false) BigDecimal price,
                                   @RequestParam(required = true) Integer rank) {
-        SalesMode salesMode =salesModeService.update(id,name,price,rank);
+        SalesMode salesMode = salesModeService.update(id, name, price, rank);
         return Result.success().msg("销售方式修改成功!").data(salesMode);
     }
 
@@ -187,7 +188,6 @@ public class CategoryController extends BaseController {
     }
 
 
-
     /**
      * 修改段位
      **/
@@ -200,11 +200,9 @@ public class CategoryController extends BaseController {
     }
 
 
-
-
-
     /**
      * 删除段位
+     *
      * @param id
      * @return
      */
